+++ data/examples/cluster/run.sh
@@ -3,3 +3,6 @@ seqcluster cluster -a seqs_map.bam -m seqs_set.ma -g ann_reduced.gtf -o test_out
 seqcluster report -j test_out_cluster/seqcluster.json -o test_out_report -r ../../genomes/genome.fa
 
 # rm -rf predict; seqcluster predict -j res/seqcluster.json --bed res/positions.bed -o test_predict -r ../genomes/genome.fa --coral --bam seqs_map_rmlw.bam
+
+rm -rf seqs_map.bam.bai
+rm -rf seqs_map_rmlw*