+++ data/examples/function/run.sh
@@ -1,1 +1,1 @@
-seqcluster target --input test_input.txt  --sps hsa --annotation ~/soft/bcbio/genomes/Hsapiens/hg19/srnaseq -o res
+seqcluster target --input test_input.txt  --sps hsa --annotation ~/soft/bcbio/genomes/Hsapiens/hg19/srnaseq -o test_out_res